moved most battle functionality into the service
